implemented UserIOConsoleImpl. fixed UserIO return types

+++ Coding/src/com/mthree/academy/c458/team3/SeanMeaney/ClassesAndObjects/ClassRoster/ui/UserIO.java
@@ -15,13 +15,13 @@ public interface UserIO {
 
     float readFloat(String prompt,float min, float max);
 
-    float readInt(String prompt);
+    int readInt(String prompt);
 
-    float readInt(String prompt,int min, int max);
+    int readInt(String prompt,int min, int max);
 
-    float readLong(String prompt);
+    long readLong(String prompt);
 
-    float readLong(String prompt,long min, long max);
+    long readLong(String prompt,long min, long max);
 
 
 }
